+++ test/runtests.jl
@@ -6,6 +6,8 @@ using Test
     include("test_solver.jl")
     include("test_connectors.jl")
     include("test_measure.jl")
+    include("test_init.jl")
     include("test_integration.jl")
+    include("test_crossings.jl")
     include("test_annotation_algorithm.jl")
 end